backend/requirements.txt
fastapi==0.109.2
uvicorn[standard]==0.27.1
pydantic==2.6.1
pydantic-settings==2.1.0
python-multipart==0.0.9
httpx==0.25.2
anthropic==0.18.1
stripe==8.4.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
PyJWT==2.8.0
email-validator==2.1.0
pytest==8.0.0
pytest-asyncio==0.23.5
supabase==2.3.4
